typing changes
