--- public/app/types/alerts.ts
@@ -1,11 +1,0 @@
-export interface AppNotification {
-  id: number;
-  severity: string;
-  icon: string;
-  title: string;
-  text: string;
-}
-
-export interface AppNotificationsState {
-  appNotifications: AppNotification[];
-}
